transaction correction

+++ foreign-remittance-impl/src/main/java/com/mislbd/ababil/foreignremittance/repository/jpa/ShadowTransactionRecordRepository.java
@@ -1,6 +1,7 @@
 package com.mislbd.ababil.foreignremittance.repository.jpa;
 
 import com.mislbd.ababil.foreignremittance.repository.schema.ShadowTransactionRecordEntity;
+import java.math.BigDecimal;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
@@ -8,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 public interface ShadowTransactionRecordRepository
     extends JpaRepository<ShadowTransactionRecordEntity, Long>, JpaSpecificationExecutor {
 
-  Optional<ShadowTransactionRecordEntity> findByGlobalTxnNo(Long globalTransactionNo);
+  Optional<ShadowTransactionRecordEntity> findByGlobalTxnNo(BigDecimal globalTransactionNo);
 }
